XWIKI-15944: Add ability to pass test resources to Docker-based tests
  * Put the method for mounting files in AbstractContainerExecutor
  * Create a dedicated class for BrowserContainerExecutor
