+++ xwiki-platform-core/xwiki-platform-test/xwiki-platform-test-docker/src/main/java/org/xwiki/test/docker/junit5/AbstractContainerExecutor.java
@@ -23,6 +23,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.output.Slf4jLogConsumer;
+import org.testcontainers.utility.MountableFile;
 
 /**
  * Common code for container execution.
@@ -50,5 +51,26 @@ public abstract class AbstractContainerExecutor
 
         // Display logs after the container has been started so that we can see problems happening in the containers
         DockerTestUtils.followOutput(container, getClass());
+    }
+
+    /**
+     * Utility method to mount a directory from the host to the container with some optimization for different OS.
+     * @param container the container on which to mount the directory.
+     * @param sourceDirectory the path of the directory on the host to mount in the container.
+     * @param targetDirectory the target where to mount the directory in the container.
+     * @since 10.11RC1
+     */
+    protected static void mountFromHostToContainer(GenericContainer container, String sourceDirectory,
+        String targetDirectory)
+    {
+        // File mounting is awfully slow on Mac OSX. For example starting Tomcat with XWiki mounted takes
+        // 45s+, while doing a COPY first and then starting Tomcat takes 8s (+5s for the copy).
+        String osName = System.getProperty("os.name").toLowerCase();
+        if (osName.startsWith("mac os x")) {
+            MountableFile mountableDirectory = MountableFile.forHostPath(sourceDirectory);
+            container.withCopyFileToContainer(mountableDirectory, targetDirectory);
+        } else {
+            container.withFileSystemBind(sourceDirectory, targetDirectory);
+        }
     }
 }
