+++ xwiki-platform-core/xwiki-platform-test/xwiki-platform-test-docker/src/main/java/org/xwiki/test/docker/junit5/TestConfiguration.java
@@ -73,8 +73,6 @@ public class TestConfiguration
 
     private static final String PROFILES_PROPERTY = "xwiki.test.ui.profiles";
 
-    private static final String TEST_RESOURCES_PATH = "/tmp/test-resources";
-
     private UITest uiTestAnnotation;
 
     private Browser browser;
@@ -460,14 +458,5 @@ public class TestConfiguration
             outputDirectory = mavenBuildDir;
         }
         return outputDirectory;
-    }
-
-    /**
-     * @return the path of the directory containing the test resources on the container.
-     * @since 10.11RC1
-     */
-    public String getTestResourcesPath()
-    {
-        return TEST_RESOURCES_PATH;
     }
 }
